Add new functionalities to chesspeek

+++ Chess/ChessUtilities/ChessUtilities/src/ChessUtilities/Board/BoardTypes.hpp
@@ -13,12 +13,10 @@ namespace chess
 {
 
 using PieceTypes = std::vector<PieceType>;
-using CoordinateDataType = int32_t;
+using CoordinateDataType = int8_t;
 using Coordinate = AlbaXY<CoordinateDataType>;
 using Coordinates = std::vector<Coordinate>;
-using Move = std::pair<Coordinate, Coordinate>;
-using Moves = std::vector<Move>;
+using Move = std::pair<Coordinate, Coordinate>;using Moves = std::vector<Move>;
 
 }
-
 }
